requirements.txt
PyYAML>=6.0
fastapi
uvicorn
pydantic
akshare
